UNICEF plugin

+++ src/addon/mod/choice/components/index/index.ts
@@ -1,4 +1,4 @@
-// (C) Copyright 2015 Moodle Pty Ltd.
+// (C) Copyright 2015 Martin Dougiamas
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
 
 import { Component, Optional, Injector } from '@angular/core';
 import { Content } from 'ionic-angular';
-import { CoreTimeUtilsProvider } from '@providers/utils/time';
 import { CoreCourseModuleMainActivityComponent } from '@core/course/classes/main-activity-component';
-import { AddonModChoiceProvider, AddonModChoiceChoice, AddonModChoiceOption, AddonModChoiceResult } from '../../providers/choice';
+import { AddonModChoiceProvider } from '../../providers/choice';
 import { AddonModChoiceOfflineProvider } from '../../providers/offline';
 import { AddonModChoiceSyncProvider } from '../../providers/sync';
+import * as moment from 'moment';
 
 /**
  * Component that displays a choice.
@@ -31,9 +31,9 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     component = AddonModChoiceProvider.COMPONENT;
     moduleName = 'choice';
 
-    choice: AddonModChoiceChoice;
-    options: AddonModChoiceOption[] = [];
-    selectedOption: {id: number};
+    choice: any;
+    options = [];
+    selectedOption: any;
     choiceNotOpenYet = false;
     choiceClosed = false;
     canEdit = false;
@@ -43,8 +43,6 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     labels = [];
     results = [];
     publishInfo: string; // Message explaining the user what will happen with his choices.
-    openTimeReadable: string;
-    closeTimeReadable: string;
 
     protected userId: number;
     protected syncEventName = AddonModChoiceSyncProvider.AUTO_SYNCED;
@@ -52,8 +50,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     protected now: number;
 
     constructor(injector: Injector, private choiceProvider: AddonModChoiceProvider, @Optional() content: Content,
-            private choiceOffline: AddonModChoiceOfflineProvider, private choiceSync: AddonModChoiceSyncProvider,
-            private timeUtils: CoreTimeUtilsProvider) {
+            private choiceOffline: AddonModChoiceOfflineProvider, private choiceSync: AddonModChoiceSyncProvider) {
         super(injector, content);
     }
 
@@ -69,8 +66,8 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
             if (!this.choice) {
                 return;
             }
-            this.choiceProvider.logView(this.choice.id, this.choice.name).then(() => {
-                this.courseProvider.checkModuleCompletion(this.courseId, this.module.completiondata);
+            this.choiceProvider.logView(this.choice.id).then(() => {
+                this.courseProvider.checkModuleCompletion(this.courseId, this.module.completionstatus);
             }).catch((error) => {
                 // Ignore errors.
             });
@@ -80,7 +77,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Perform the invalidate content function.
      *
-     * @return Resolved when done.
+     * @return {Promise<any>} Resolved when done.
      */
     protected invalidateContent(): Promise<any> {
         const promises = [];
@@ -98,8 +95,8 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Compares sync event data with current data to check if refresh content is needed.
      *
-     * @param syncEventData Data receiven on sync observer.
-     * @return True if refresh is needed, false otherwise.
+     * @param {any} syncEventData Data receiven on sync observer.
+     * @return {boolean}          True if refresh is needed, false otherwise.
      */
     protected isRefreshSyncNeeded(syncEventData: any): boolean {
         if (this.choice && syncEventData.choiceId == this.choice.id && syncEventData.userId == this.userId) {
@@ -114,22 +111,22 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Download choice contents.
      *
-     * @param refresh If it's refreshing content.
-     * @param sync If it should try to sync.
-     * @param showErrors If show errors to the user of hide them.
-     * @return Promise resolved when done.
+     * @param  {boolean}      [refresh=false]    If it's refreshing content.
+     * @param  {boolean}      [sync=false]       If the refresh is needs syncing.
+     * @param  {boolean}      [showErrors=false] If show errors to the user of hide them.
+     * @return {Promise<any>} Promise resolved when done.
      */
     protected fetchContent(refresh: boolean = false, sync: boolean = false, showErrors: boolean = false): Promise<any> {
         this.now = new Date().getTime();
 
         return this.choiceProvider.getChoice(this.courseId, this.module.id).then((choice) => {
             this.choice = choice;
-            this.choice.timeopen = choice.timeopen * 1000;
-            this.choice.timeclose = choice.timeclose * 1000;
-            this.openTimeReadable = this.timeUtils.userDate(choice.timeopen);
-            this.closeTimeReadable = this.timeUtils.userDate(choice.timeclose);
+            this.choice.timeopen = parseInt(choice.timeopen) * 1000;
+            this.choice.openTimeReadable = moment(choice.timeopen).format('LLL');
+            this.choice.timeclose = parseInt(choice.timeclose) * 1000;
+            this.choice.closeTimeReadable = moment(choice.timeclose).format('LLL');
 
-            this.description = choice.intro;
+            this.description = choice.intro || choice.description;
             this.choiceNotOpenYet = choice.timeopen && choice.timeopen > this.now;
             this.choiceClosed = choice.timeclose && choice.timeclose <= this.now;
 
@@ -156,7 +153,8 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
             return this.fetchOptions(hasOffline).then(() => {
                 return this.fetchResults();
             });
-        }).finally(() => {
+        }).then(() => {
+            // All data obtained, now fill the context menu.
             this.fillContextMenu(refresh);
         });
     }
@@ -164,8 +162,8 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Convenience function to get choice options.
      *
-     * @param hasOffline True if there are responses stored offline.
-     * @return Promise resolved when done.
+     * @param {boolean} hasOffline True if there are responses stored offline.
+     * @return {Promise<any>} Promise resolved when done.
      */
     protected fetchOptions(hasOffline: boolean): Promise<any> {
         return this.choiceProvider.getOptions(this.choice.id).then((options) => {
@@ -176,7 +174,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
 
             if (hasOffline) {
                 promise = this.choiceOffline.getResponse(this.choice.id).then((response) => {
-                    const optionsKeys: {[id: number]: AddonModChoiceOption} = {};
+                    const optionsKeys = {};
                     options.forEach((option) => {
                         optionsKeys[option.id] = option;
                     });
@@ -224,7 +222,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
                 promise = Promise.resolve(options);
             }
 
-            promise.then((options: AddonModChoiceOption[]) => {
+            promise.then((options) => {
                 const isOpen = this.isChoiceOpen();
 
                 let hasAnswered = false;
@@ -278,7 +276,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Convenience function to get choice results.
      *
-     * @return Resolved when done.
+     * @return {Promise<any>} Resolved when done.
      */
     protected fetchResults(): Promise<any> {
         if (this.choiceNotOpenYet) {
@@ -292,11 +290,11 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
             let hasVotes = false;
             this.data = [];
             this.labels = [];
-            results.forEach((result: AddonModChoiceResultFormatted) => {
+            results.forEach((result) => {
                 if (result.numberofuser > 0) {
                     hasVotes = true;
                 }
-                result.percentageamountfixed = result.percentageamount.toFixed(1);
+                result.percentageamount = parseFloat(result.percentageamount).toFixed(1);
                 this.data.push(result.numberofuser);
                 this.labels.push(result.text);
             });
@@ -308,7 +306,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Check if a choice is open.
      *
-     * @return True if choice is open, false otherwise.
+     * @return {boolean} True if choice is open, false otherwise.
      */
     protected isChoiceOpen(): boolean {
         return (this.choice.timeopen === 0 || this.choice.timeopen <= this.now) &&
@@ -318,7 +316,7 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Return true if the user has selected at least one option.
      *
-     * @return True if the user has responded.
+     * @return {boolean} True if the user has responded.
      */
     canSave(): boolean {
         if (this.choice.allowmultiple) {
@@ -353,13 +351,14 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
             }
 
             const modal = this.domUtils.showModalLoading('core.sending', true);
-            this.choiceProvider.submitResponse(this.choice.id, this.choice.name, this.courseId, responses).then((online) => {
+            this.choiceProvider.submitResponse(this.choice.id, this.choice.name, this.courseId, responses).then(() => {
                 // Success!
                 // Check completion since it could be configured to complete once the user answers the choice.
-                this.courseProvider.checkModuleCompletion(this.courseId, this.module.completiondata);
+                this.courseProvider.checkModuleCompletion(this.courseId, this.module.completionstatus);
                 this.domUtils.scrollToTop(this.content);
 
-                return this.dataUpdated(online);
+                // Let's refresh the data.
+                return this.refreshContent(false);
             }).catch((message) => {
                 this.domUtils.showErrorModalDefault(message, 'addon.mod_choice.cannotsubmit', true);
             }).finally(() => {
@@ -372,12 +371,12 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
      * Delete options selected.
      */
     delete(): void {
-        this.domUtils.showDeleteConfirm().then(() => {
+        this.domUtils.showConfirm(this.translate.instant('core.areyousure')).then(() => {
             const modal = this.domUtils.showModalLoading('core.sending', true);
             this.choiceProvider.deleteResponses(this.choice.id, this.choice.name, this.courseId).then(() => {
                 this.domUtils.scrollToTop(this.content);
 
-                // Refresh the data. Don't call dataUpdated because deleting an answer doesn't mark the choice as outdated.
+                // Success! Let's refresh the data.
                 return this.refreshContent(false);
             }).catch((message) => {
                 this.domUtils.showErrorModalDefault(message, 'addon.mod_choice.cannotsubmit', true);
@@ -390,31 +389,9 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     }
 
     /**
-     * Function to call when some data has changed. It will refresh/prefetch data.
-     *
-     * @param online Whether the data was sent to server or stored in offline.
-     * @return Promise resolved when done.
-     */
-    protected dataUpdated(online: boolean): Promise<any> {
-        if (online && this.isPrefetched()) {
-            // The choice is downloaded, update the data.
-            return this.choiceSync.prefetchAfterUpdate(this.module, this.courseId).then(() => {
-                // Update the view.
-                this.showLoadingAndFetch(false, false);
-            }).catch(() => {
-                // Prefetch failed, refresh the data.
-                return this.refreshContent(false);
-            });
-        } else {
-            // Not downloaded, refresh the data.
-            return this.refreshContent(false);
-        }
-    }
-
-    /**
      * Performs the sync of the activity.
      *
-     * @return Promise resolved when done.
+     * @return {Promise<any>} Promise resolved when done.
      */
     protected sync(): Promise<any> {
         return this.choiceSync.syncChoice(this.choice.id, this.userId);
@@ -423,17 +400,10 @@ export class AddonModChoiceIndexComponent extends CoreCourseModuleMainActivityCo
     /**
      * Checks if sync has succeed from result sync data.
      *
-     * @param result Data returned on the sync function.
-     * @return Whether it succeed or not.
+     * @param  {any} result Data returned on the sync function.
+     * @return {boolean} Whether it succeed or not.
      */
     protected hasSyncSucceed(result: any): boolean {
         return result.updated;
     }
 }
-
-/**
- * Choice result with some calculated data.
- */
-export type AddonModChoiceResultFormatted = AddonModChoiceResult & {
-    percentageamountfixed: string; // Percentage of users answers with fixed decimals.
-};
