+++ src/addon/mod/forum/components/post/post.ts
@@ -1,4 +1,4 @@
-// (C) Copyright 2015 Moodle Pty Ltd.
+// (C) Copyright 2015 Martin Dougiamas
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,25 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {
-    Component, Input, Output, Optional, EventEmitter, OnInit, OnDestroy, ViewChild, ElementRef, OnChanges, SimpleChange
-} from '@angular/core';
+import { Component, Input, Output, Optional, EventEmitter, OnInit, OnDestroy } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { Content, PopoverController, ModalController } from 'ionic-angular';
+import { NavController } from 'ionic-angular';
 import { TranslateService } from '@ngx-translate/core';
 import { CoreFileUploaderProvider } from '@core/fileuploader/providers/fileuploader';
+import { CoreSplitViewComponent } from '@components/split-view/split-view';
 import { CoreSyncProvider } from '@providers/sync';
 import { CoreDomUtilsProvider } from '@providers/utils/dom';
 import { CoreTextUtilsProvider } from '@providers/utils/text';
-import { CoreEventsProvider } from '@providers/events';
-import { CoreSitesProvider } from '@providers/sites';
 import { AddonModForumProvider } from '../../providers/forum';
 import { AddonModForumHelperProvider } from '../../providers/helper';
 import { AddonModForumOfflineProvider } from '../../providers/offline';
 import { AddonModForumSyncProvider } from '../../providers/sync';
-import { CoreRatingInfo } from '@core/rating/providers/rating';
-import { CoreTagProvider } from '@core/tag/providers/tag';
-import { AddonForumPostOptionsMenuComponent } from '../post-options-menu/post-options-menu';
 
 /**
  * Components that shows a discussion post, its attachments and the action buttons allowed (reply, etc.).
@@ -39,7 +33,8 @@ import { AddonForumPostOptionsMenuComponent } from '../post-options-menu/post-op
     selector: 'addon-mod-forum-post',
     templateUrl: 'addon-mod-forum-post.html',
 })
-export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges {
+export class AddonModForumPostComponent implements OnInit, OnDestroy {
+
     @Input() post: any; // Post.
     @Input() courseId: number; // Post's course ID.
     @Input() discussionId: number; // Post's' discussion ID.
@@ -49,26 +44,17 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
     @Input() originalData: any; // Object with the original post data. Usually shared between posts.
     @Input() trackPosts: boolean; // True if post is being tracked.
     @Input() forum: any; // The forum the post belongs to. Required for attachments and offline posts.
-    @Input() accessInfo: any; // Forum access information.
-    @Input() parentSubject?: string; // Subject of parent post.
-    @Input() ratingInfo?: CoreRatingInfo; // Rating info item.
-    @Input() leavingPage?: boolean; // Whether the page that contains this post is being left and will be destroyed.
+    @Input() defaultSubject: string; // Default subject to set to new posts.
     @Output() onPostChange: EventEmitter<void>; // Event emitted when a reply is posted or modified.
-
-    @ViewChild('replyFormEl') formElement: ElementRef;
 
     messageControl = new FormControl();
 
     uniqueId: string;
-    defaultReplySubject: string;
-    advanced = false; // Display all form fields.
-    tagsEnabled: boolean;
-    displaySubject = true;
-    optionsMenuEnabled = false;
 
     protected syncId: string;
 
     constructor(
+            private navCtrl: NavController,
             private uploaderProvider: CoreFileUploaderProvider,
             private syncProvider: CoreSyncProvider,
             private domUtils: CoreDomUtilsProvider,
@@ -78,14 +64,8 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
             private forumHelper: AddonModForumHelperProvider,
             private forumOffline: AddonModForumOfflineProvider,
             private forumSync: AddonModForumSyncProvider,
-            private tagProvider: CoreTagProvider,
-            @Optional() private content: Content,
-            protected popoverCtrl: PopoverController,
-            protected modalCtrl: ModalController,
-            protected eventsProvider: CoreEventsProvider,
-            protected sitesProvider: CoreSitesProvider) {
+            @Optional() private svComponent: CoreSplitViewComponent) {
         this.onPostChange = new EventEmitter<void>();
-        this.tagsEnabled = this.tagProvider.areTagsAvailableInSite();
     }
 
     /**
@@ -93,80 +73,37 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
      */
     ngOnInit(): void {
         this.uniqueId = this.post.id ? 'reply' + this.post.id : 'edit' + this.post.parent;
-
-        const reTranslated = this.translate.instant('addon.mod_forum.re');
-        this.displaySubject = !this.parentSubject ||
-            (this.post.subject != this.parentSubject && this.post.subject != `Re: ${this.parentSubject}` &&
-                this.post.subject != `${reTranslated} ${this.parentSubject}`);
-        this.defaultReplySubject = (this.post.subject.startsWith('Re: ') || this.post.subject.startsWith(reTranslated))
-            ? this.post.subject : `${reTranslated} ${this.post.subject}`;
-
-        this.optionsMenuEnabled = !this.post.id || (this.forumProvider.isGetDiscussionPostAvailable() &&
-                    (this.forumProvider.isDeletePostAvailable() || this.forumProvider.isUpdatePostAvailable()));
     }
 
     /**
-     * Detect changes on input properties.
-     */
-    ngOnChanges(changes: {[name: string]: SimpleChange}): void {
-        if (changes.leavingPage && this.leavingPage) {
-            // Download all courses is enabled now, initialize it.
-            this.domUtils.triggerFormCancelledEvent(this.formElement, this.sitesProvider.getCurrentSiteId());
-        }
-    }
-
-    /**
-     * Deletes an online post.
-     */
-    deletePost(): void {
-        this.domUtils.showDeleteConfirm('addon.mod_forum.deletesure').then(() => {
-            const modal = this.domUtils.showModalLoading('core.deleting', true);
-
-            this.forumProvider.deletePost(this.post.id).then((response) => {
-
-                const data = {
-                    forumId: this.forum.id,
-                    discussionId: this.discussionId,
-                    cmId: this.forum.cmid,
-                    deleted: response.status,
-                    post: this.post
-                };
-
-                this.eventsProvider.trigger(AddonModForumProvider.CHANGE_DISCUSSION_EVENT, data,
-                    this.sitesProvider.getCurrentSiteId());
-
-                this.domUtils.showToast('addon.mod_forum.deletedpost', true);
-            }).catch((error) => {
-                this.domUtils.showErrorModal(error);
-            }).finally(() => {
-                modal.dismiss();
-            });
-        }).catch((error) => {
-            // Do nothing.
-        });
-    }
-
-    /**
-     * Set data to new reply post, clearing temporary files and updating original data.
+     * Opens the profile of a user.
      *
-     * @param replyingTo Id of post beeing replied.
-     * @param isEditing True it's an offline reply beeing edited, false otherwise.
-     * @param subject Subject of the reply.
-     * @param message Message of the reply.
-     * @param isPrivate True if it's private reply.
-     * @param files Reply attachments.
+     * @param {number} userId
      */
-    protected setReplyFormData(replyingTo?: number, isEditing?: boolean, subject?: string, message?: string, files?: any[],
-            isPrivate?: boolean): void {
+    openUserProfile(userId: number): void {
+        // Decide which navCtrl to use. If this page is inside a split view, use the split view's master nav.
+        const navCtrl = this.svComponent ? this.svComponent.getMasterNav() : this.navCtrl;
+        navCtrl.push('CoreUserProfilePage', {userId, courseId: this.courseId});
+    }
+
+    /**
+     * Set data to new post, clearing temporary files and updating original data.
+     *
+     * @param {number} [replyingTo] Id of post beeing replied.
+     * @param {boolean} [isEditing] True it's an offline reply beeing edited, false otherwise.
+     * @param {string} [subject] Subject of the reply.
+     * @param {string} [message] Message of the reply.
+     * @param {any[]} [files] Reply attachments.
+     */
+    protected setReplyData(replyingTo?: number, isEditing?: boolean, subject?: string, message?: string, files?: any[]): void {
         // Delete the local files from the tmp folder if any.
         this.uploaderProvider.clearTmpFiles(this.replyData.files);
 
         this.replyData.replyingTo = replyingTo || 0;
         this.replyData.isEditing = !!isEditing;
-        this.replyData.subject = subject || this.defaultReplySubject || '';
+        this.replyData.subject = subject || this.defaultSubject || '';
         this.replyData.message = message || null;
         this.replyData.files = files || [];
-        this.replyData.isprivatereply = !!isPrivate;
 
         // Update rich text editor.
         this.messageControl.setValue(this.replyData.message);
@@ -175,161 +112,39 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
         this.originalData.subject = this.replyData.subject;
         this.originalData.message = this.replyData.message;
         this.originalData.files = this.replyData.files.slice();
-        this.originalData.isprivatereply = this.replyData.isprivatereply;
-
-        // Show advanced fields if any of them has not the default value.
-        this.advanced = this.replyData.files.length > 0;
-    }
-
-    /**
-     * Show the context menu.
-     *
-     * @param e Click Event.
-     */
-    showOptionsMenu(e: Event): void {
-        e.preventDefault();
-        e.stopPropagation();
-
-        const popover = this.popoverCtrl.create(AddonForumPostOptionsMenuComponent, {
-            post: this.post,
-            forumId: this.forum.id
-        });
-        popover.onDidDismiss((data) => {
-            if (data && data.action) {
-                switch (data.action) {
-                    case 'edit':
-                        this.editPost();
-                        break;
-                    case 'editoffline':
-                        this.editOfflineReply();
-                        break;
-                    case 'delete':
-                        this.deletePost();
-                        break;
-                    case 'deleteoffline':
-                        this.discardOfflineReply();
-                        break;
-                    default:
-                        break;
-                }
-            }
-        });
-        popover.present({
-            ev: e
-        });
-    }
-
-    /**
-     * Shows a form modal to edit an online post.
-     */
-    editPost(): void {
-        const modal = this.modalCtrl.create('AddonModForumEditPostPage', {
-            post: this.post,
-            component: this.component,
-            componentId:  this.componentId,
-            forum: this.forum
-        });
-
-        modal.present();
-        modal.onDidDismiss((data) => {
-            if (typeof data != 'undefined') {
-                // Add some HTML to the message if needed.
-                const message = this.textUtils.formatHtmlLines(data.message);
-                const files = data.files || [];
-                const sendingModal = this.domUtils.showModalLoading('core.sending', true);
-                let promise;
-
-                // Upload attachments first if any.
-                if (files.length) {
-                    promise = this.forumHelper.uploadOrStoreReplyFiles(this.forum.id, this.post.id, files, false);
-                } else {
-                    promise = Promise.resolve();
-                }
-
-                promise.then((attach) => {
-                    const options: any = {};
-
-                    if (attach) {
-                        options.attachmentsid = attach;
-                    }
-
-                    // Try to send it to server.
-                    return this.forumProvider.updatePost(this.post.id, data.subject, message, options);
-                }).then((sent) => {
-                    if (sent && this.forum.id) {
-                        // Data sent to server, delete stored files (if any).
-                        this.forumHelper.deleteReplyStoredFiles(this.forum.id, this.post.id);
-
-                        this.onPostChange.emit();
-                        this.post.subject = data.subject;
-                        this.post.message = message;
-                        this.post.attachments = data.files;
-                    }
-                }).catch((message) => {
-                    this.domUtils.showErrorModalDefault(message, 'addon.mod_forum.couldnotupdate', true);
-                }).finally(() => {
-                    sendingModal.dismiss();
-                });
-            }
-        });
     }
 
     /**
      * Set this post as being replied to.
      */
-    showReplyForm(): void {
+    showReply(): void {
         if (this.replyData.isEditing) {
             // User is editing a post, data needs to be resetted. Ask confirm if there is unsaved data.
             this.confirmDiscard().then(() => {
-                this.setReplyFormData(this.post.id);
-
-                if (this.content) {
-                    setTimeout(() => {
-                        this.content.resize();
-                        this.domUtils.scrollToElementBySelector(this.content, '#addon-forum-reply-edit-form-' + this.uniqueId);
-                    });
-                }
+                this.setReplyData(this.post.id);
             }).catch(() => {
                 // Cancelled.
             });
-
-            return;
         } else if (!this.replyData.replyingTo) {
             // User isn't replying, it's a brand new reply. Initialize the data.
-            this.setReplyFormData(this.post.id);
+            this.setReplyData(this.post.id);
         } else {
             // The post being replied has changed but the data will be kept.
             this.replyData.replyingTo = this.post.id;
-
-            if (this.replyData.subject == this.originalData.subject) {
-                // Update subject only if it hadn't been modified
-                this.replyData.subject = this.defaultReplySubject;
-                this.originalData.subject = this.defaultReplySubject;
-            }
-
             this.messageControl.setValue(this.replyData.message);
         }
-
-        if (this.content) {
-            setTimeout(() => {
-                this.content.resize();
-                this.domUtils.scrollToElementBySelector(this.content, '#addon-forum-reply-edit-form-' + this.uniqueId);
-            });
-        }
-
     }
 
     /**
      * Set this post as being edited to.
      */
-    editOfflineReply(): void {
+    editReply(): void {
         // Ask confirm if there is unsaved data.
         this.confirmDiscard().then(() => {
             this.syncId = this.forumSync.getDiscussionSyncId(this.discussionId);
             this.syncProvider.blockOperation(AddonModForumProvider.COMPONENT, this.syncId);
 
-            this.setReplyFormData(this.post.parent, true, this.post.subject, this.post.message, this.post.attachments,
-                    this.post.isprivatereply);
+            this.setReplyData(this.post.parent, true, this.post.subject, this.post.message, this.post.attachments);
         }).catch(() => {
             // Cancelled.
         });
@@ -338,7 +153,7 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
     /**
      * Message changed.
      *
-     * @param text The new text.
+     * @param {string} text The new text.
      */
     onMessageChange(text: string): void {
         this.replyData.message = text;
@@ -371,11 +186,6 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
 
         // Add some HTML to the message if needed.
         message = this.textUtils.formatHtmlLines(message);
-
-        // Set private option if checked.
-        if (this.replyData.isprivatereply) {
-            options.private = true;
-        }
 
         // Upload attachments first if any.
         if (files.length) {
@@ -419,11 +229,9 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
             }
 
             // Reset data.
-            this.setReplyFormData();
+            this.setReplyData();
 
             this.onPostChange.emit();
-
-            this.domUtils.triggerFormSubmittedEvent(this.formElement, sent, this.sitesProvider.getCurrentSiteId());
 
             if (this.syncId) {
                 this.syncProvider.unblockOperation(AddonModForumProvider.COMPONENT, this.syncId);
@@ -441,9 +249,7 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
     cancel(): void {
         this.confirmDiscard().then(() => {
             // Reset data.
-            this.setReplyFormData();
-
-            this.domUtils.triggerFormCancelledEvent(this.formElement, this.sitesProvider.getCurrentSiteId());
+            this.setReplyData();
 
             if (this.syncId) {
                 this.syncProvider.unblockOperation(AddonModForumProvider.COMPONENT, this.syncId);
@@ -456,8 +262,8 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
     /**
      * Discard offline reply.
      */
-    discardOfflineReply(): void {
-        this.domUtils.showDeleteConfirm().then(() => {
+    discard(): void {
+        this.domUtils.showConfirm(this.translate.instant('core.areyousure')).then(() => {
             const promises = [];
 
             promises.push(this.forumOffline.deleteReply(this.post.parent));
@@ -469,7 +275,7 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
 
             return Promise.all(promises).finally(() => {
                 // Reset data.
-                this.setReplyFormData();
+                this.setReplyData();
 
                 this.onPostChange.emit();
 
@@ -480,20 +286,6 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
         }).catch(() => {
             // Cancelled.
         });
-    }
-
-    /**
-     * Function called when rating is updated online.
-     */
-    ratingUpdated(): void {
-        this.forumProvider.invalidateDiscussionPosts(this.discussionId, this.forum.id);
-    }
-
-    /**
-     * Show or hide advanced form fields.
-     */
-    toggleAdvanced(): void {
-        this.advanced = !this.advanced;
     }
 
     /**
@@ -508,7 +300,7 @@ export class AddonModForumPostComponent implements OnInit, OnDestroy, OnChanges 
     /**
      * Confirm discard changes if any.
      *
-     * @return Promise resolved if the user confirms or data was not changed and rejected otherwise.
+     * @return {Promise<void>} Promise resolved if the user confirms or data was not changed and rejected otherwise.
      */
     protected confirmDiscard(): Promise<void> {
         if (this.forumHelper.hasPostDataChanged(this.replyData, this.originalData)) {
